Removed PARALLEL_HACK code
Removed PARALLEL_HACK related code as I realized I could achieve what I
wanted via the Apache configuration (I was almost fully there anyway).  I
did lose the ability to flip back and forth on the single link, but I felt
that the removal of the code for *one entry* was more than worth it, and the
entry in question still gets the point across.