Added the Post To Facebook Feature
It took a bit longer than I expected, but I wanted to do it right.

Anyway, new config options:

Facebook-AP-ID:		Facebook Appliation ID
Facebook-AP-Secret:	Facebook Application Secret
Facebook-User:		Facebook user to update

Also, added a status header (which is saved with each post now) and
added it to the template system.

I also added a facebook flag for email and the web interface.

Also found a bug in generate_pages() in that it trashes the req structure.
I'll have to work on that later.