More reworking of how the blog entries are working.
I did away with the cache.  On reflection, I don't think it every worked
properly to begin with, and with the heavy file caching that Linux does
anyway, we rarely, if ever, hit the physical disk (and honestly, at this
point, I doubt this will ever run on anything other than Linux).  So go
for simplicity and obviously working than trying to be clever.

This does, however, muck with the writing of entries.