A bit about comments
The blog engine has some support for comments, but very rudementary support.
I mean, you can display some, and if you have an appropriate script that
can generate (and/or modify) the appropriate file, you too, can have comments.

But really, I never bothered, and because of spam, it's quite a bit of
trouble, and everybody is using FaceGoogleBookPlus for comments *anyway* so
this is of some dubious value ...