Bug fix---cb_cond_blog_title()
cb_cond_blog_title() was being given a struct callback_data, not a List (but
that was okay because the first element of a struct callback_data *is* a
list), but it wasn't checking to see if entry was valid---it just assumed it
was.  The assert that was there was triggering.  Removed to an actual test.

The entry is not valid if you give an entry that doesn't exist.