Bug fix---tumbler problems, but not necessarily with the parsing
The intent was if the start date was later than the end date, swap the two
dates and set the reverse flag.  Unfortunately, the logic to determine the
swap was wrong.  Wong wrong wrong wrong wrong.  I moved the check down, and
rewrote the check logic.  It should be fine now, but get back to me in
another decade.