Bug fix---double free
Memory managament is a bitch if you aren't careful about object lifetimes.
cmd_cgi_post_show() was calling BlogEntryFree(), but the entry in question
had already been freed in cb_entry() as part of displaying the page.  I
think this was the last bit of code at this level that actually freed
entries (all others at this level have been removed).

This is what happens in seldom used code---bug accumlate in the dark corners
and fester like this until an emergency happens and BOOM!  You end up not
being able to post to the blog during a power outtage.