Bug fix---encoded entries via email mess things up.
Some entries have been received encoded as 'quoted-printable'.  I'm still
unsure of why that happens, but it does, and it really messes things up as
I'm not expecting that.

The bug fix checks for '7bit', '8bit' and 'binary' encodings (they're
identity encodings---these terms are descriptive) and if not one of those,
stop processing and signal an error.