Document a crashing bug.
If a request such as "GET /boston.cgi HTTP/1.1" happens, the program (as it
was) crashed with an assert from empty_string().  Changing the call site (in
cmd_cgi_get_show()) to emptynull_string() now causes the program to
regenerate the cached pages and send nothing to the request.

What should happen is probably a 400 error.  Given my current schedule on
vacation, it make take me a few days to get this addressed.