All bugs being tracked here have been fixed
As well as other bugs never entered into the file.  Any remaining bugs (or
possible problems) are marked in the code with XXX, so let's try tracking
the bugs in the code itself.  It may be a bad idea, but we won't know until
we try.

And Lord knows I've had enough bad ideas.