Finish up the new tumbler work.
For the first time since this started, the reverse tumbler code is probably
correct!  And redirection on ranges works!  I think I finally got the
tumbler stuff working correctly!

Woot!

It's about the same amount of code, and probably just as inscrutable to the
neophyte (or me after fifteen years) as ever.  Hopefully not---it's more
straightforward than the previous version.

All that's left is some more testing, and some clean up here and there.