Normalize const and restrict usage on codebase.
Parameters that I could mark as const, I did.  Pointers that are declared of
the form of:

	int foo(ptr_t *const a)

mean a single pointer of that type, not an array of pointers.  It has no
effect on the code generated, but it does signal programmer intent.

The restrict qualifier is only used on parameter lists, and only if two or
more pointers of the same type are being passed in---again to signify
programmer intent that the two objects are distinct, not potential aliased
objects.