Bug fix---getline() called with incorrect parameters.
It just happened to work.  line and size should be set to NULL and 0
respectively, but they weren't.  It just so happened that the stack hadn't
been used at that point in the execution, and because Linux uses address 0
as NULL, it just was never triggered.

That wasn't the root cause of the lastest crash (it was due to misspelling
the author header) but it is a bug I found and one that should be fixed.