Bug fix---I was a bit too assertive in the callback code.
There were some asserts that shouldn't have been there.  But they did bring
up a few bugs that just happened not to happen.  When serving x-html, it
attempts to write the file with the template of the blog, but the normal
data is NULL.  This isn't that bad of a case since the code already works
around it, but the assert for NULL triggered the assertion that data wasn't
NULL when it could be.

I need to think on this.