A constant refactoring
The language spec for C states that 'const' modified the thing to its right,
*except* if it appears first in a declaration, in which case, it modifies
the thing on the left.  I refactored the code such that all the 'const'
modifiers appeared on the right of what it modifies, which is more
consistent across the board.

I shall expect to become used to this in time.