Bug fix---potential overflow when allocating memory
When allocating memory with the sequence:

	x = malloc(n * sizeof(type));

you risk the multiplication overflowing.  To check for that, you can check
if n is less than:

	(SIZE_MAX - sizeof(type)) + 1) / sizeof(type)

Example:  If SIZE_MAX is 65535, and type is int (2 bytes), this comes down
to:

	(65535 - 2) + 1) / 2
	(65533      + 1) / 2
	 65534           / 2
 	 32767

So if n in this case is larger than 32767, we can't safely allocate any
memory.