The 'Completely Arbitrary Christmas Release' Version
I'm moving away from semantic versioning [1] for this, as I feel that
semantic versioning doesn't really work for applications---it's good for
libraries, but applications feel different.  They are less reused and more
just built upon.  A single ever-incrementing version number is probably
good, but I do release bug-only fixes, so for me, a dual version X.Y is
nice---X for any non-but release, and .Y for purely bug fixes.

Works for me.