Add hooks to blogging engine.
Much like VCS have hooks to run programs at certain points, I just added the
feature to mod_blog.  There are currently two hooks---one for when posts are
added, and one for when a template is run, both are post-hooks.

The hook for entries receives a single argument, the URL of the post.
The hook for templates receives a single argument, the filename.

These hooks are not meant to stop processing, but are really intended to
allow an external program to notify other systems of the added entry (think
webping.com from years ago).