Bug fix---Apache 2.4 fix
Apache 2.4 changed how PATH_TRANSLATED is handled in the face of
redirections---it borks the string by just appending "redirect:" to
PATH_INFO and does NOT translate the string.  So change the code to do the
right thing in the face of redirections, and one that will still work on
older versions of Apache.