Remove the final three global constant variables.
I moved them to the config structure, but I need to work on getting ride of
them.  One of them I don't use, and I don't think anyone else uses as I
don't think anyone else uses mod_blog to run a blog.  The other two deal
with URL stuff and that's hard enough as it is.