Cut update option
The update option was to set some metadata that I've long since forgotten
about and don't use.  I also don't think anyone else but me uses this
blogging engine, so screw it!  Thus begins some major cleanup of the code,
and the removal of other optional code that I don't use.  Given the lack of
pull requests, or issues on Github, I don't think I have to worry about
this.