Remove the conf.h file.
Since I'm the only user of this, and the conf.h file was configured for my
use, just set those options in stone, and remove the need for it entirely.
This also has the benefit of removing code that hasn't really been tested
since it was written 20 years ago under a different set of
assumptions---namely, that I woudn't be the only one running this.