Initialize the global display structure
Since I have to initialize gd.template in code, I figure any settings I did
have in the pre-inialized struct could be moved to code as well.  It keeps
the initialization in one place, it decreases the size of the data segment,
and it's not that much code being added.