Refactor the config code yet again.
This time, I made the config reading code more isolated, so it can be easier
to reuse in other projects (much like I did for the tumblers).  This
involved making the conversion and pagegen fields strings, and writing code
to convert the string to actual function pointers elsewhere.  The code is
getting in better shape, but I'm still struggling to reduce the global
variables.  I still feel like it's cheating putting all the global variables
into a strcture---I'm not fooling anyone with that.