Major reworking of callbacks.c
Reordered the functions such that all the static fuctions appear before
they're used, so we can remove all the prototypes.  Also, reorder the
callback functions in alphabetical order.  If you have that many, some order
is better than just random.

Also, move the struct chunk_callback structure to generic_cb()---no need
for that to be visible to the entire file.