Wheee!  I get to make bloody noisy diffs!
Tighten up the code around a struct.  Yes, doing this aligning of data makes
for some very noisy diffs, but I don't care!  I'm into the aesthetics of the
code, and yes, that does make me unsual.  I bend the code formatting to *my*
whims, not to the whims of a bloody formatter with no taste on how the code
should be formatted, but was written to make it easy to write the formatter.