Add support files for webmentions.
Not all support is included here.  Missing is a program to accept
webmentions.  See <https://www.w3.org/TR/webmention/> for more information.
This program should generate a file of "<entry#>.webmention" where each
webmention is on a line.  The URL of the mention comes first, then an ASCII
TAB character, followed by text, and then end of line.