Mark arrays with [].
I had a "char ***plines".  Arg!  Okay, let's break that down:  a *char is a
string, so we can assume we have:

	typedef char *string;

So "char ***plines" can be thought of as "string **plines".  A bit better.
But also, the declaration of "char **" is (most often) an array of strings,
so a better way to write that is "string foo[]".  So that gets rid of two
stars.  All we're left with is now "string *foo[]", which (usually) means
the callee is going to allocate the memory for the array of strings.  Now,
I'm not the type of person to use the typedef above (because 'char *' is too
idiomatic for C to change now).  So instead, I have:

	char **plines[]

which is a bit better than what I had before.  Also, I went ahead and found
other instances of a pointer var that is actually an array and marked them
as such in the parameter list.  Intent signalling and all that.